requirements.txt
asgiref==3.8.1
Django==5.2.2
django-debug-toolbar==5.2.0
django-htmx==1.23.0
django-widget-tweaks==1.5.0
python-dotenv==1.1.0
sqlparse==0.5.3
